move http_backport to its own file, fix python 2 linting
